% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelDataGuide-package.R
\docType{package}
\name{excelDataGuide-package}
\alias{excelDataGuide}
\alias{excelDataGuide-package}
\title{excelDataGuide: Read Data from templated Excel Data Reports}
\description{
Provides functions as well as a YAML guide format to facilitate reading data from filled-out excel templates. In particular, the package provides such facilities for experiments performed with microplates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://systemsbioinformatics.github.io/excelDataGuide/}
  \item Report bugs at \url{https://github.com/douwe/excelDataGuide/issues}
}

}
\author{
\strong{Maintainer}: Douwe Molenaar \email{d.molenaar@vu.nl} (\href{https://orcid.org/0000-0001-7108-4545}{ORCID})

Authors:
\itemize{
  \item Douwe Molenaar \email{d.molenaar@vu.nl} (\href{https://orcid.org/0000-0001-7108-4545}{ORCID})
}

}
\keyword{internal}
