% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wells.R
\name{format_wells}
\alias{format_wells}
\title{Standardize a vector with well names.}
\usage{
format_wells(v, format)
}
\arguments{
\item{v}{A vector with potentially sloppy well names}

\item{format}{A single element character or numeric vector with the format of the plate}
}
\value{
A vector with standardized well names
}
\description{
Standardize a vector with well names.
}
\examples{
format_wells(c("a01", "A 2", "0", " A 4 ", "A05", "H012"), 96)


}
