% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_star}
\alias{has_star}
\title{Check if a string has a star}
\usage{
has_star(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A logical vector
}
\description{
This function checks if a string has a star, x or a question mark.
}
