% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kvlist_to_table}
\alias{kvlist_to_table}
\title{Create a table from a list of key-value pairs}
\usage{
kvlist_to_table(kvlist, guide, reverse.translate = TRUE)
}
\arguments{
\item{kvlist}{A list of key-value pairs}

\item{guide}{A data guide}

\item{reverse.translate}{A logical indicating whether to reverse translate the keys}
}
\value{
A data frame with columns 'key' and 'value'
}
\description{
Create a table from a list of key-value pairs
}
