% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{long_to_shortnames}
\alias{long_to_shortnames}
\title{Translate long variable names to short variable names}
\usage{
long_to_shortnames(v, translations)
}
\arguments{
\item{v}{A vector of long variable names}

\item{translations}{A named vector with long variable names as names and short variable names as values}
}
\value{
A vector of short variable names
}
\description{
Translate long variable names to short variable names
}
