% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wells.R
\name{normalize_wells}
\alias{normalize_wells}
\title{Normalize a vector with well names.}
\usage{
normalize_wells(v, format = NULL)
}
\arguments{
\item{v}{A vector with potentially sloppy well names}

\item{format}{A single element character or numeric vector with the format of the plate, or NULL}
}
\value{
A vector with normalized well names
}
\description{
Normalized well names are in the format \code{A1}, \code{B2}, \code{H12}, etc., i.e., an
uppercase letter followed by an integer without zero-padding and without
spaces. If the well name can not be converted to a normalized value an NA is
returned. If the plate format is given then the well names are checked to see
if they are present in the format, and are converted to NA if not.
}
\examples{
normalize_wells(c("a01", "A 2", "0", " A 4 ", "A05", "H012", "K12"), 96)
normalize_wells(c("a01", "A 2", "0", " A 4 ", "A05", "H012", "K12"))

}
