% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wells.R
\name{parse_well_series}
\alias{parse_well_series}
\title{Try to parse a string with a potential series of wells}
\usage{
parse_well_series(v, format)
}
\arguments{
\item{v}{A character vector of length 1 with a potential series of well names}

\item{format}{A well format string (e.g., "96", "384", "1536")}
}
\value{
A vector with well names or \code{NA} values
}
\description{
Users sometimes enter multiple well names in a single string. This function
attempts to parse such a string and returns a vector of well names.
}
