% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wells.R
\name{rowcol_from_well}
\alias{rowcol_from_well}
\title{Calculate row and column from well name.}
\usage{
rowcol_from_well(well, format)
}
\arguments{
\item{well}{A character vector with the standard well names}

\item{format}{A single element character or numeric vector with the format of the plate}
}
\value{
A list with two elements: row and col
}
\description{
Calculate row and column from well name.
}
\examples{
rowcol_from_well(c("A1", "B2", "C3", NA), 48)
# The order is preserved
rowcol_from_well(c("H12", "A1"), 96)
}
