% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{short_to_longnames}
\alias{short_to_longnames}
\title{Reverse translate short variable names to long variable names}
\usage{
short_to_longnames(v, translations)
}
\arguments{
\item{v}{A vector of long variable names}

\item{translations}{A named vector with long variable names as names and short variable names as values}
}
\value{
A vector of long variable names
}
\description{
Reverse translate short variable names to long variable names
}
