% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{star_to_number}
\alias{star_to_number}
\title{Convert a string with a star to a number}
\usage{
star_to_number(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A numeric vector
}
\description{
If a template uses the convention of putting a star in front of numbers that
should be excluded then this function can be used to convert the variable to
a number again.
}
