% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{long_to_shortnames}
\alias{long_to_shortnames}
\alias{short_to_longnames}
\title{Translation between long and short variable names}
\usage{
long_to_shortnames(v, translations)

short_to_longnames(v, translations)
}
\arguments{
\item{v}{A vector of variable names}

\item{translations}{A table of translations with columns \code{long} and \code{short}}
}
\value{
A vector of long or short variable names

A vector of long variable names
}
\description{
Translate between long and short variable names. If a translation is missing the original
variable long or short variable name from \code{v} is returned.
}
