% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide.R
\name{read_guide}
\alias{read_guide}
\title{Read a spreadsheet data guide}
\usage{
read_guide(path)
}
\arguments{
\item{path}{The path to the guide file}
}
\value{
A list object with the guide data
}
\description{
A spreadsheet guide is a YAML file that contains directions to where data
can be found in a spreadsheet. For an extensive description of this file
type, as well as instructions how to create and test the validity, please
see the vignettes.
}
