% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wells.R
\name{well_from_rowcol}
\alias{well_from_rowcol}
\title{Calculate well number from row and column. Function used in \code{raw_map} function
from package \code{platetools}.}
\usage{
well_from_rowcol(row, col)
}
\arguments{
\item{row}{A character vector with the row names}

\item{col}{A numeric vector with the column names}
}
\value{
A character vector with the well numbers
}
\description{
Calculate well number from row and column. Function used in \code{raw_map} function
from package \code{platetools}.
}
